Use static setLifecycle instead of setter for testing
